Refactor: Improve code organization and structure
- Moved function definitions to appropriate source files
- Added header files with function declarations
- Ensured single definition of global variables
- Improved file documentation and comments
