+++ STM32/src/config.c
@@ -1,0 +1,72 @@
+#include "config.h"
+
+void system_clock_setup(void)
+{
+    rcc_clock_setup_pll(&rcc_hse_configs[RCC_CLOCK_HSE8_72MHZ]);
+    rcc_periph_clock_enable(RCC_AFIO);
+}
+
+void gpio_setup(void)
+{
+    rcc_periph_clock_enable(RCC_GPIOA);
+    gpio_set_mode(GPIOA, GPIO_MODE_INPUT, GPIO_CNF_INPUT_ANALOG, GPIO0);
+}
+
+void uart_setup(void)
+{
+    rcc_periph_clock_enable(RCC_GPIOA);
+    rcc_periph_clock_enable(RCC_USART2);
+
+    gpio_set_mode(GPIOA, GPIO_MODE_OUTPUT_50_MHZ, GPIO_CNF_OUTPUT_ALTFN_PUSHPULL, GPIO_USART2_TX);
+    gpio_set_mode(GPIOA, GPIO_MODE_INPUT, GPIO_CNF_INPUT_FLOAT, GPIO_USART2_RX);
+
+    usart_set_baudrate(USART2, 9600);
+    usart_set_databits(USART2, 8);
+    usart_set_stopbits(USART2, USART_STOPBITS_1);
+    usart_set_mode(USART2, USART_MODE_TX_RX);
+    usart_set_parity(USART2, USART_PARITY_NONE);
+    usart_set_flow_control(USART2, USART_FLOWCONTROL_NONE);
+
+    usart_enable(USART2);
+    usart_enable_rx_interrupt(USART2);
+    nvic_enable_irq(NVIC_USART2_IRQ);
+}
+
+void adc_setup(void)
+{
+    rcc_periph_clock_enable(RCC_ADC1);
+    rcc_periph_clock_enable(RCC_DMA1);
+
+    adc_power_off(ADC1);
+    adc_enable_scan_mode(ADC1);
+    adc_set_single_conversion_mode(ADC1);
+
+    adc_set_sample_time(ADC1, ADC_CHANNEL0, ADC_SMPR_SMP_239DOT5CYC);
+    uint8_t channel_array[] = {ADC_CHANNEL0};
+    adc_set_regular_sequence(ADC1, 1, channel_array);
+
+    adc_enable_dma(ADC1);
+    adc_power_on(ADC1);
+    adc_reset_calibration(ADC1);
+    adc_calibrate(ADC1);
+
+    dma_channel_reset(DMA1, DMA_CHANNEL1);
+    dma_set_peripheral_address(DMA1, DMA_CHANNEL1, (uint32_t)&ADC_DR(ADC1));
+    dma_set_memory_address(DMA1, DMA_CHANNEL1, (uint32_t)&adc_value);
+    dma_set_number_of_data(DMA1, DMA_CHANNEL1, 1);
+    dma_set_read_from_peripheral(DMA1, DMA_CHANNEL1);
+
+    dma_set_peripheral_size(DMA1, DMA_CHANNEL1, DMA_CCR_PSIZE_32BIT);
+    dma_set_memory_size(DMA1, DMA_CHANNEL1, DMA_CCR_MSIZE_32BIT);
+
+    dma_disable_memory_increment_mode(DMA1, DMA_CHANNEL1);
+    dma_enable_memory_increment_mode(DMA1, DMA_CHANNEL1);
+    dma_set_priority(DMA1, DMA_CHANNEL1, DMA_CCR_PL_HIGH);
+
+    dma_enable_transfer_complete_interrupt(DMA1, DMA_CHANNEL1);
+    dma_enable_transfer_error_interrupt(DMA1, DMA_CHANNEL1);
+    nvic_enable_irq(NVIC_DMA1_CHANNEL1_IRQ);
+
+    dma_enable_channel(DMA1, DMA_CHANNEL1);
+    adc_start_conversion_direct(ADC1);
+}
